refactor(routes): align lookup-person with top-level handler pattern
- convert lookup-person route registration from class-based controller to top-level routing functions
- replace Main registration with handleLookupPerson(openSearchClient)
- rename LookupPersonController.kt to LookupPerson.kt for naming consistency
- keep endpoint behavior unchanged and verify with compile + lookup-person tests
